Broke out the js debug methods into their own helper module.
